+++ src/modules/changelog.md
@@ -4,11 +4,11 @@
   - Added tests
   - Added `valueOf` for DataBus which refers to the last taken data
   - Added `Warden.Formula`
-  - Added `Warden.From`
   - Added configureable Data Bus values history (3 by default)
   - Renamed internal modules `DataBus` to `Stream`
   - Added `.stream` method to Streams objects, now `Bus` is deprecated
   - Fixed reduce method
+  - Fixed `Stream.data.last` reference
    
 `version: v.0.3.0`